js_hw#1

+++ js/task-3.js
@@ -1,6 +1,6 @@
 function getElementWidth(content, padding, border) {
-    const contentWidth = Number.parseInt(content);
-    const paddingWidth = Number.parseInt(padding);
+    const contentWidth = Number.parseFloat(content);
+    const paddingWidth = Number.parseFloat(padding);
     const borderWidth = Number.parseFloat(border);
 
     return contentWidth + (paddingWidth * 2) + (borderWidth * 2);
